backend/requirements.txt
# FastAPI dependencies (kept for reference - backend now uses Django)
# fastapi==0.121.1
# uvicorn==0.38.0
# aiosqlite==0.21.0
# python-multipart==0.0.20
# pydantic==2.12.4

# Django REST Framework dependencies
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

